Add files via upload
This report is created as a one-stop view for a Senior Executive showing multiple stats and metrics reviewed at weekly cadence. The dashboards are created on Tableau desktop and published on company's Tableau server. The data is sensitive and confidential information and is pushed here on my private Github only to showcase my Tableau analysis and dashboarding skills.