Add files via upload
This file shows Engagement ring intended Qualified visitors to the website and funnel of their shopping journey. In addition, it also shows other sales stats of these customers.
Report is used by Executive team on a weekly cadence. 
Data is sensitive and confidential in this report and is only intended to be used to judge my analytics and dashboarding skills on Tableau.